fix console.warnings
