+++ packages/nextra/src/client/components/index.ts
@@ -15,4 +15,4 @@ export { Tr } from './tr.js'
 export { Mermaid } from '@theguild/remark-mermaid/mermaid'
 export { MathJax, MathJaxContext } from 'better-react-mathjax'
 export { Playground } from './playground.js'
-export { Popup, PopupTrigger, PopupPanel } from './popup.js'
+export { Popup } from './popup.js'
